Extract the test runners for `node:child_process`
Extract the function that "run a test" against functions from the
`node:child_process` API from the `test/fuzz` folder into a separate
module that is more clearly intended for sharing (between e2e tests and
fuzzing).

Also included some minor refactoring to the code being changed, most
notably reducing duplication in the e2e test macros.
