+++ test/fuzz/fork.test.cjs
@@ -4,49 +4,18 @@
  * @license MIT
  */
 
-const assert = require("node:assert");
-const { fork } = require("node:child_process");
-
-const common = require("./_common.cjs");
-
-const shescape = require("../../index.cjs");
-
-function check(arg) {
-  const forkOptions = { silent: true };
-
-  const safeArg = shescape.escape(arg);
-
-  return new Promise((resolve, reject) => {
-    const echo = fork(common.ECHO_SCRIPT, [safeArg], forkOptions);
-
-    echo.on("error", (error) => {
-      reject(`an unexpected error occurred: ${error}`);
-    });
-
-    echo.stdout.on("data", (data) => {
-      const result = data.toString();
-      const expected = common.getExpectedOutput({ arg });
-      try {
-        assert.strictEqual(result, expected);
-        resolve();
-      } catch (e) {
-        reject(e);
-      }
-    });
-  });
-}
+const runners = require("../e2e/_runners.cjs");
 
 async function fuzz(buf) {
   const arg = buf.toString();
 
   try {
-    await check(arg);
+    await runners.fork(arg);
   } catch (e) {
     throw e;
   }
 }
 
 module.exports = {
-  check,
   fuzz,
 };
